Use flatten=full option
