+++ go/models/audit_log_payload.go
@@ -10,7 +10,6 @@ import (
 
 	strfmt "github.com/go-openapi/strfmt"
 
-	"github.com/go-openapi/errors"
 	"github.com/go-openapi/swag"
 )
 
@@ -178,11 +177,6 @@ func (m AuditLogPayload) MarshalJSON() ([]byte, error) {
 
 // Validate validates this audit log payload
 func (m *AuditLogPayload) Validate(formats strfmt.Registry) error {
-	var res []error
-
-	if len(res) > 0 {
-		return errors.CompositeValidationError(res...)
-	}
 	return nil
 }
