+++ go/models/payment_method.go
@@ -87,40 +87,3 @@ func (m *PaymentMethod) UnmarshalBinary(b []byte) error {
 	*m = res
 	return nil
 }
-
-// PaymentMethodData payment method data
-// swagger:model PaymentMethodData
-type PaymentMethodData struct {
-
-	// card type
-	CardType string `json:"card_type,omitempty"`
-
-	// email
-	Email string `json:"email,omitempty"`
-
-	// last4
-	Last4 string `json:"last4,omitempty"`
-}
-
-// Validate validates this payment method data
-func (m *PaymentMethodData) Validate(formats strfmt.Registry) error {
-	return nil
-}
-
-// MarshalBinary interface implementation
-func (m *PaymentMethodData) MarshalBinary() ([]byte, error) {
-	if m == nil {
-		return nil, nil
-	}
-	return swag.WriteJSON(m)
-}
-
-// UnmarshalBinary interface implementation
-func (m *PaymentMethodData) UnmarshalBinary(b []byte) error {
-	var res PaymentMethodData
-	if err := swag.ReadJSON(b, &res); err != nil {
-		return err
-	}
-	*m = res
-	return nil
-}
