+++ go/models/site_setup.go
@@ -17,8 +17,7 @@ import (
 type SiteSetup struct {
 	Site
 
-	// repo
-	Repo *RepoInfo `json:"repo,omitempty"`
+	SiteSetupAllOf1
 }
 
 // UnmarshalJSON unmarshals this object from a JSON structure
@@ -31,14 +30,11 @@ func (m *SiteSetup) UnmarshalJSON(raw []byte) error {
 	m.Site = aO0
 
 	// AO1
-	var dataAO1 struct {
-		Repo *RepoInfo `json:"repo,omitempty"`
-	}
-	if err := swag.ReadJSON(raw, &dataAO1); err != nil {
+	var aO1 SiteSetupAllOf1
+	if err := swag.ReadJSON(raw, &aO1); err != nil {
 		return err
 	}
-
-	m.Repo = dataAO1.Repo
+	m.SiteSetupAllOf1 = aO1
 
 	return nil
 }
@@ -53,17 +49,11 @@ func (m SiteSetup) MarshalJSON() ([]byte, error) {
 	}
 	_parts = append(_parts, aO0)
 
-	var dataAO1 struct {
-		Repo *RepoInfo `json:"repo,omitempty"`
+	aO1, err := swag.WriteJSON(m.SiteSetupAllOf1)
+	if err != nil {
+		return nil, err
 	}
-
-	dataAO1.Repo = m.Repo
-
-	jsonDataAO1, errAO1 := swag.WriteJSON(dataAO1)
-	if errAO1 != nil {
-		return nil, errAO1
-	}
-	_parts = append(_parts, jsonDataAO1)
+	_parts = append(_parts, aO1)
 
 	return swag.ConcatJSON(_parts...), nil
 }
@@ -76,32 +66,14 @@ func (m *SiteSetup) Validate(formats strfmt.Registry) error {
 	if err := m.Site.Validate(formats); err != nil {
 		res = append(res, err)
 	}
-
-	if err := m.validateRepo(formats); err != nil {
+	// validation for a type composition with SiteSetupAllOf1
+	if err := m.SiteSetupAllOf1.Validate(formats); err != nil {
 		res = append(res, err)
 	}
 
 	if len(res) > 0 {
 		return errors.CompositeValidationError(res...)
 	}
-	return nil
-}
-
-func (m *SiteSetup) validateRepo(formats strfmt.Registry) error {
-
-	if swag.IsZero(m.Repo) { // not required
-		return nil
-	}
-
-	if m.Repo != nil {
-		if err := m.Repo.Validate(formats); err != nil {
-			if ve, ok := err.(*errors.Validation); ok {
-				return ve.ValidateName("repo")
-			}
-			return err
-		}
-	}
-
 	return nil
 }
